🐛 fix: 올바르지 않은 동작(버그)을 고친 경우
폴더명 대소문자 -> 소문자로 변경. 이에 맞춰서 import 구문도 소문자로 변경함.
변경된 사항을 github 원격 레포지토리에도 적용하도록 gitconfig 세팅

break: npm run gitconfig 로 원격 레포지토리가 대소문자 구별할 수 있도록 프로젝트 팀원들도 세팅할 것.
