+++ .github/workflows/lint-check.yml
@@ -28,4 +28,5 @@ jobs:
         run: npm ci
 
       - name: Run prettier and eslint
-        run: npm run prettier-lint
+        run: |
+          npm run prettier-lint || (echo "You should run prettier-lint before pr" && exit 1)
