+++ src/components/ui/atoms/no-question-box/NoQuestionBox.jsx
